backend/requirements.txt
fastapi>=0.115.0
uvicorn==0.30.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
python-dotenv==1.0.1
starlette>=0.47.2
python-multipart>=0.0.22
orjson>=3.11.6
